+++ cmd/server/setup.go
@@ -56,8 +56,8 @@ func setupStore(ctx context.Context, c *cli.Command) (store.Store, error) {
 	xorm := store.XORM{
 		Log:             c.Bool("db-log"),
 		ShowSQL:         c.Bool("db-log-sql"),
-		MaxOpenConns:    int(c.Int64("db-max-open-connections")),
-		MaxIdleConns:    int(c.Int64("db-max-idle-connections")),
+		MaxOpenConns:    c.Int("db-max-open-connections"),
+		MaxIdleConns:    c.Int("db-max-idle-connections"),
 		ConnMaxLifetime: c.Duration("db-max-connection-timeout"),
 	}
